# openapi-sprint1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IQBF ULIMA API",
    "summary": "API transaccional para la gestión de IQBF",
    "description": "Contrato del Sprint 1: identidad, organización, insumos, presentaciones, densidades versionadas y auditoría.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Desarrollo local"
    }
  ],
  "paths": {
...
    "/api/insumos/{id_insumo}/presentaciones": {
      "get": {
        "tags": [
          "Insumos y presentaciones"
        ],
        "summary": "Listar las presentaciones de un insumo",
        "operationId": "list_presentations_api_insumos__id_insumo__presentaciones_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_insumo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id Insumo"
            }
          },
          {
            "name": "incluir_inactivas",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Incluir Inactivas"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PresentacionOut"
                  },
                  "title": "Response List Presentations Api Insumos  Id Insumo  Presentaciones Get"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida o inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "403": {
            "description": "Permiso denegado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "404": {
            "description": "Recurso inexistente",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto con el estado actual",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "422": {
            "description": "Validación o regla de negocio",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PresentacionOut": {
        "properties": {
          "id_presentacion": {
            "type": "string",
            "title": "Id Presentacion"
          },
          "codigo_bf_sunat": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codigo Bf Sunat"
          },
          "codigo_presentacion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codigo Presentacion"
          },
          "concentracion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concentracion"
          },
          "capacidad": {
            "type": "string",
            "pattern": "^-?\\d+(\\.\\d+)?$",
            "title": "Capacidad",
            "example": "1250.0000"
          },
          "unidad": {
            "type": "string",
            "enum": [
              "mL",
              "L",
              "g",
              "kg"
            ],
            "title": "Unidad"
          },
          "tipo_envase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tipo Envase"
          },
          "equivalencia_g": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^-?\\d+(\\.\\d+)?$",
                "example": "1250.0000"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equivalencia G"
          },
          "densidad_actual": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^-?\\d+(\\.\\d+)?$",
                "example": "1250.0000"
              },
              {
                "type": "null"
              }
            ],
            "title": "Densidad Actual"
          },
          "estado": {
            "type": "string",
            "enum": [
              "VIGENTE",
              "INACTIVO"
            ],
            "title": "Estado"
          },
          "id_insumo": {
            "type": "string",
            "title": "Id Insumo"
          },
          "vigencia_desde": {
            "type": "string",
            "format": "date",
            "title": "Vigencia Desde"
          },
          "vigencia_hasta": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vigencia Hasta"
          }
        },
        "type": "object",
        "required": [
          "id_presentacion",
          "codigo_bf_sunat",
          "codigo_presentacion",
          "concentracion",
          "capacidad",
          "unidad",
          "tipo_envase",
          "equivalencia_g",
          "densidad_actual",
          "estado",
          "id_insumo",
          "vigencia_desde",
          "vigencia_hasta"
        ],
        "title": "PresentacionOut"
      },
...
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "instance": {
            "type": "string",
            "title": "Instance"
          },
          "code": {
            "type": "string",
            "enum": [
              "VALIDACION",
              "ERROR_INTERNO",
              "ERROR_BASE_DATOS",
              "REGISTRO_DUPLICADO",
              "REFERENCIA_INVALIDA",
              "REGLA_NEGOCIO",
              "SALDO_INSUFICIENTE",
              "KARDEX_INMUTABLE",
              "SALDO_SOLO_VIA_KARDEX",
              "DENSIDAD_REQUERIDA",
              "VIGENCIA_SOLAPADA",
              "SESION_EXPIRADA",
              "TOKEN_INVALIDO",
              "NO_AUTENTICADO",
              "SESION_INVALIDA",
              "CREDENCIALES_INVALIDAS",
              "CUENTA_INACTIVA",
              "CUENTA_BLOQUEADA",
              "SIN_ROL",
              "PERMISO_DENEGADO",
              "CATALOGO_NO_EXISTE",
              "CATALOGO_SOLO_LECTURA",
              "REGISTRO_NO_EXISTE",
              "ESTABLECIMIENTO_REQUERIDO",
              "CODIGO_INSTITUCIONAL_REQUERIDO",
              "INSUMO_NO_EXISTE",
              "PRESENTACION_NO_EXISTE",
              "VIGENCIA_INVALIDA",
              "USUARIO_NO_EXISTE",
              "ROL_REQUERIDO",
              "ROL_INVALIDO",
              "ALCANCE_INVALIDO",
              "ALCANCE_GLOBAL_REQUERIDO",
              "MOTIVO_REQUERIDO",
              "DENSIDAD_LOTE_REQUERIDA",
              "INSUMO_INACTIVO"
            ],
            "title": "Code"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProblemFieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "request_id"
        ],
        "title": "Problem"
      },
      "ProblemFieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "field",
          "message",
          "type"
        ],
        "title": "ProblemFieldError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}